Bugfix causing panels to have a wrong width/height
FollowObjectBehaviour is a behaviour that's moving it's container to the same position as the owner of the container. This means it should update the position of the container when:
	1. parent moves
	2. relative properties of container changes
	3. size of the container changes

When the behaviour updates the position directly after 2&3, it will cause the container to have a wrong width/height. If the behaviour waits one frame, there's no bug?!